+++ volunteerus-server/src/questions/questions.controller.ts
@@ -1,17 +1,26 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete } from '@nestjs/common';
+import { Controller, Get, Post, Body, Patch, Param, Delete, Query } from '@nestjs/common';
 import { QuestionsService } from './questions.service';
 import { CreateQuestionDto } from './dto/create-question.dto';
 import { UpdateQuestionDto } from './dto/update-question.dto';
 import { Question } from './schemas/question.schema';
 import mongoose from 'mongoose';
+import { CaslAbilityFactory } from 'src/casl/casl-ability.factory/casl-ability.factory';
+
 
 @Controller('questions')
 export class QuestionsController {
-  constructor(private readonly questionsService: QuestionsService) {}
+  constructor(
+    private readonly questionsService: QuestionsService,
+    private caslAbilityFactory: CaslAbilityFactory
+  ) {}
 
   @Post()
-  create(@Body() createQuestionDto: CreateQuestionDto): Promise<Question> {
-    return this.questionsService.create(createQuestionDto);
+  create(@Query('role') role: string, @Body() createQuestionDto: CreateQuestionDto): Promise<Question> {
+    // Check if user has permission to create event details / sign up form
+    const ability = this.caslAbilityFactory.createForUser(role);
+    if (ability.can('create', Question)) {
+      return this.questionsService.create(createQuestionDto);
+    } 
   }
 
   @Get()
@@ -25,13 +34,21 @@ export class QuestionsController {
   }
 
   @Patch(':id')
-  update(@Param('id') id: mongoose.Types.ObjectId, @Body() updateQuestionDto: UpdateQuestionDto) {
-    return this.questionsService.update(id, updateQuestionDto);
+  update(@Param('id') id: mongoose.Types.ObjectId, @Query('role') role: string, @Body() updateQuestionDto: UpdateQuestionDto) {
+    // Check if user has permission to update event details / sign up form
+    const ability = this.caslAbilityFactory.createForUser(role);
+    if (ability.can('update', Question)) {
+      return this.questionsService.update(id, updateQuestionDto);
+    }
   }
 
   @Delete(':id')
-  remove(@Param('id') id: mongoose.Types.ObjectId): Promise<Question> {
-    return this.questionsService.remove(id);
+  remove(@Query('role') role: string, @Param('id') id: mongoose.Types.ObjectId): Promise<Question> {
+    // Check if user has permission to delete event details / sign up form
+    const ability = this.caslAbilityFactory.createForUser(role);
+    if (ability.can('delete', Question)) {
+      return this.questionsService.remove(id);
+    }
   }
 
 }
